rename component to match class name
